requirements.txt
langgraph-supervisor
google-auth 
google-auth-oauthlib 
google-api-python-client 
pandas 
langchain_google_genai
langchain
langgraph
flask
langchain_groq
docs